apps/sylvia-app/requirements.txt
# Web/API
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==2.3.0

# GUI & visualization
matplotlib==3.8.2

# YAML storage for custom training data
pyyaml==6.0

# Core ML / data libraries used by personality engine
numpy==1.26.0
pandas==2.1.2
scikit-learn==1.3.0

# Optional: HTTP requests for API or microservices
requests==2.31.0

# Optional: LLM API integration if using OpenAI
# openai==0.27.8